Top Down Recursive Parser (for Verb Only)
The Purpose of this Parser is to analyze using TOP-DOWN RECURSIVE TREE
technique and return if the entered Sanskrit word is a verb or not. It
checks the word by checking its vibhakti i.e. Suffix, and in case of
Past Tense words even its prefix. For more efficiency and accuracy
recognition of Dhaatu can be done by creating a database for dhaatus.
Note: In creating such a database ambiguities will be there that needs
to handled properly.