Sanskrit Parser using DFA
The Purpose of this Parser is to analyze and return if the entered
Sanskrit Sentence is Gramatically Correct or not,ie in SOV(Subject
Object Verb Form)or not. It checks the word by checking its vibhakti
i.e. Suffix, and in some cases prefix too. For more efficiency and
accuracy recognition of Dhaatu can be done by creating a database for
dhaatus. Note: In creating such a database ambiguities will be there
that needs to handled properly.
It will work only for a Three word input.